requirements.txt
torch>=1.10
torchtext==0.4.0
spacy>=3.7
scikit-learn>=1.4
matplotlib>=3.8
numpy>=1.26
seaborn>=0.13
PyYAML>=6.0
pandas>=2.1
